# [FIX] account: auto_join to speedup _kanban_dashboard