@Aurelienvd 2021-09-10
@odoo/rd-accounting Could someone take a look at this PR when he/she has some time ?