@william-andre 2021-09-13
https://github.com/odoo/odoo/pull/66924